requirements.txt
regex
tweepy
python-decouple